requirements.txt
fastapi>=0.85.0
uvicorn>=0.19.0
google-cloud-firestore>=2.7.1
python-multipart>=0.0.5
pdfplumber>=0.7.4
python-docx>=0.8.11
openai>=0.27.8
python-dotenv>=0.21.0
tiktoken>=0.4.0